Implement divide_horizontal_then_vertical_with_weights
