+++ src/area.rs
@@ -1,0 +1,4 @@
+/// Area of an axis aligned rectangle
+pub trait Area<T> {
+    fn area(&self) -> T;
+}
