+++ src/aspect_ratio.rs
@@ -1,0 +1,3 @@
+pub(crate) trait AspectRatio<T> {
+    fn aspect_ratio(&self) -> T;
+}
