+++ src/dividing.rs
@@ -1,6 +1,7 @@
 use crate::{
+    area::Area,
     axis::{Axis, SizeForAxis},
-    rectangle::{Area, RectangleSize},
+    rectangle::RectangleSize,
     rotate::QuarterRotation,
 };
 
@@ -53,7 +54,10 @@ pub trait Dividing<T> {
             + std::ops::Div<Output = T>
             + std::ops::Mul<Output = T>,
     {
-        if weights.len() <= 1 {
+        if weights.is_empty() {
+            return vec![];
+        }
+        if weights.len() == 1 {
             return vec![self.clone()];
         }
         let sum_of_weights: T = weights.iter().sum();
@@ -118,6 +122,29 @@ pub trait Dividing<T> {
 
         divided
     }
+
+    fn divide_horizontal_then_vertical_with_weights(
+        &self,
+        weights: &[T],
+        aspect_ratio: T,
+    ) -> Vec<Self>
+    where
+        Self: Sized + RectangleSize<T> + Clone + SizeForAxis<T> + Area<T> + QuarterRotation,
+        T: Copy
+            + std::ops::Add<Output = T>
+            + for<'a> std::iter::Sum<&'a T>
+            + std::ops::Div<Output = T>
+            + std::ops::Mul<Output = T>
+            + std::cmp::PartialOrd,
+    {
+        // rotate, divide vertical, rotate back again means divide horizontal
+        let rotated = self.rotate_clockwise();
+        let divided = rotated.divide_vertical_then_horizontal_with_weights(weights, aspect_ratio);
+        divided
+            .iter()
+            .map(|r| r.rotate_counter_clockwise())
+            .collect()
+    }
 }
 
 pub(crate) trait VerticalDividingHelper<T> {
@@ -147,9 +174,111 @@ where
 mod tests {
     use super::*;
     use crate::axis_aligned_rectangle::AxisAlignedRectangle;
+    use crate::component::Component;
     use crate::point::Point;
     use crate::rectangle::Rectangle;
     use crate::weight::normalize_weights;
+
+    #[test]
+    fn test_divide_vertical() {
+        let point = Point::new(2, 3);
+        let rect = Rectangle::new(4, 5);
+        let (rect_a, rect_b) = AxisAlignedRectangle::new(point, rect).divide_vertical(2);
+        assert_eq!(rect_a.origin(), point);
+        assert_eq!(rect_a.rect(), Rectangle::new(2, 5));
+        assert_eq!(rect_b.origin(), Point::new(4, 3));
+        assert_eq!(rect_b.rect(), Rectangle::new(2, 5));
+
+        let point = Point::new(2, 3);
+        let rect = Rectangle::new(4, 5);
+        let (rect_a, rect_b) = AxisAlignedRectangle::new(point, rect).divide_vertical(1);
+        assert_eq!(rect_a.origin(), point);
+        assert_eq!(rect_a.rect(), Rectangle::new(1, 5));
+        assert_eq!(rect_b.origin(), Point::new(3, 3));
+        assert_eq!(rect_b.rect(), Rectangle::new(3, 5));
+    }
+
+    #[test]
+    fn test_divide_horizontal() {
+        let point = Point::new(2, 3);
+        let rect = Rectangle::new(4, 5);
+        let (rect_a, rect_b) = AxisAlignedRectangle::new(point, rect).divide_horizontal(1);
+        assert_eq!(rect_a.origin(), point);
+        assert_eq!(rect_a.rect(), Rectangle::new(4, 1));
+        assert_eq!(rect_b.origin(), Point::new(2, 4));
+        assert_eq!(rect_b.rect(), Rectangle::new(4, 4));
+
+        let point = Point::new(2, 3);
+        let rect = Rectangle::new(4, 5);
+        let (rect_a, rect_b) = AxisAlignedRectangle::new(point, rect).divide_horizontal(2);
+        assert_eq!(rect_a.origin(), point);
+        assert_eq!(rect_a.rect(), Rectangle::new(4, 2));
+        assert_eq!(rect_b.origin(), Point::new(2, 5));
+        assert_eq!(rect_b.rect(), Rectangle::new(4, 3));
+    }
+
+    #[test]
+    fn test_divide_nth() {
+        // test vertical
+        let point = Point::new(2, 3);
+        let rect = Rectangle::new(6, 2);
+        let a_rect = AxisAlignedRectangle::new(point, rect);
+        let divided = a_rect.divide_by_values_and_axis(&vec![1, 2], Axis::Vertical);
+        assert_eq!(divided[0].origin(), point);
+        assert_eq!(divided[0].rect(), Rectangle::new(1, 2));
+        assert_eq!(divided[1].origin(), Point::new(3, 3));
+        assert_eq!(divided[1].rect(), Rectangle::new(2, 2));
+        assert_eq!(divided[2].origin(), Point::new(5, 3));
+        assert_eq!(divided[2].rect(), Rectangle::new(3, 2));
+        assert_no_overlaps(&a_rect, &divided);
+        assert_eq!(divided.len(), 3);
+        // sum of divided rectangles should equal original rectangle
+        assert_eq!(
+            divided[0].width() + divided[1].width() + divided[2].width(),
+            a_rect.width()
+        );
+        // all divided rectangles should have the same height
+        assert_eq!(divided[0].height(), a_rect.height());
+        assert_eq!(divided[1].height(), a_rect.height());
+        assert_eq!(divided[2].height(), a_rect.height());
+        // the sum of the x and width of the  divided rectangle should equal the x of the next divided rectangle
+        assert_eq!(divided[0].x() + divided[0].width(), divided[1].x());
+        assert_eq!(divided[1].x() + divided[1].width(), divided[2].x());
+        assert_eq!(
+            a_rect.x() + a_rect.width(),
+            divided[2].x() + divided[2].width()
+        );
+
+        // test horizontal
+        let point = Point::new(2, 3);
+        let rect = Rectangle::new(2, 6);
+        let a_rect = AxisAlignedRectangle::new(point, rect);
+        let divided = a_rect.divide_by_values_and_axis(&vec![3, 2], Axis::Horizontal);
+        assert_eq!(divided[0].origin(), point);
+        assert_eq!(divided[0].rect(), Rectangle::new(2, 3));
+        assert_eq!(divided[1].origin(), Point::new(2, 6));
+        assert_eq!(divided[1].rect(), Rectangle::new(2, 2));
+        assert_eq!(divided[2].origin(), Point::new(2, 8));
+        assert_eq!(divided[2].rect(), Rectangle::new(2, 1));
+        assert_no_overlaps(&a_rect, &divided);
+        assert_eq!(divided.len(), 3);
+        // sum of divided rectangles should equal original rectangle
+        assert_eq!(
+            divided[0].height() + divided[1].height() + divided[2].height(),
+            a_rect.height()
+        );
+        // all divided rectangles should have the same width
+        assert_eq!(divided[0].width(), a_rect.width());
+        assert_eq!(divided[1].width(), a_rect.width());
+        assert_eq!(divided[2].width(), a_rect.width());
+        // the sum of the y and height of the  divided rectangle should equal the y of the next divided rectangle
+        assert_eq!(divided[0].y() + divided[0].height(), divided[1].y());
+        assert_eq!(divided[1].y() + divided[1].height(), divided[2].y());
+        assert_eq!(
+            a_rect.y() + a_rect.height(),
+            divided[2].y() + divided[2].height()
+        );
+    }
 
     #[test]
     fn test_divide_vertical_then_horizontal_with_weights() {
@@ -173,7 +302,7 @@ mod tests {
         let weights = vec![1.0, 1.0, 1.0, 1.0];
         let divided = rect.divide_vertical_then_horizontal_with_weights(&weights, 1.0);
         assert_weights_dividing(&rect, &divided, &weights);
-        assert_weights_dividing_no_overlaps(&rect, &divided);
+        assert_no_overlaps(&rect, &divided);
         assert_eq!(
             divided[0],
             AxisAlignedRectangle::new(Point::new(0.0, 0.0), Rectangle::new(50.0, 50.0))
@@ -196,8 +325,49 @@ mod tests {
         let weights = vec![1.0];
         let divided = rect.divide_vertical_then_horizontal_with_weights(&weights, 1.0);
         assert_weights_dividing(&rect, &divided, &weights);
-        assert_weights_dividing_no_overlaps(&rect, &divided);
+        assert_no_overlaps(&rect, &divided);
         assert_eq!(divided[0], rect);
+    }
+
+    #[test]
+    fn test_divide_horizontal_then_vertical_with_weights() {
+        let rect = Rectangle::new(100.0, 100.0);
+        let weights = vec![1.0, 1.0, 1.0, 1.0];
+        let divided = rect.divide_horizontal_then_vertical_with_weights(&weights, 1.0);
+        assert_weights_dividing(&rect, &divided, &weights);
+        assert_eq!(divided[0], Rectangle::new(50.0, 50.0));
+        assert_eq!(divided[1], Rectangle::new(50.0, 50.0));
+        assert_eq!(divided[2], Rectangle::new(50.0, 50.0));
+        assert_eq!(divided[3], Rectangle::new(50.0, 50.0));
+
+        // not divided case
+        let rect = Rectangle::new(100.0, 100.0);
+        let weights = vec![1.0];
+        let divided = rect.divide_horizontal_then_vertical_with_weights(&weights, 1.0);
+        assert_weights_dividing(&rect, &divided, &weights);
+        assert_eq!(divided[0], rect);
+
+        let rect = AxisAlignedRectangle::new(Point::new(0.0, 0.0), Rectangle::new(100.0, 100.0));
+        let weights = vec![1.0, 1.0, 1.0, 1.0];
+        let divided = rect.divide_horizontal_then_vertical_with_weights(&weights, 1.0);
+        assert_weights_dividing(&rect, &divided, &weights);
+        assert_no_overlaps(&rect, &divided);
+        assert_eq!(
+            divided[0],
+            AxisAlignedRectangle::new(Point::new(0.0, 0.0), Rectangle::new(50.0, 50.0))
+        );
+        assert_eq!(
+            divided[1],
+            AxisAlignedRectangle::new(Point::new(50.0, 0.0), Rectangle::new(50.0, 50.0))
+        );
+        assert_eq!(
+            divided[2],
+            AxisAlignedRectangle::new(Point::new(0.0, 50.0), Rectangle::new(50.0, 50.0))
+        );
+        assert_eq!(
+            divided[3],
+            AxisAlignedRectangle::new(Point::new(50.0, 50.0), Rectangle::new(50.0, 50.0))
+        );
     }
 
     fn assert_weights_dividing<T, D>(original: &D, divided: &[D], weights: &[T])
@@ -227,7 +397,7 @@ mod tests {
         assert_eq!(original_normalized_weights, divided_area_by_weights);
     }
 
-    fn assert_weights_dividing_no_overlaps<T>(
+    fn assert_no_overlaps<T>(
         original: &AxisAlignedRectangle<T>,
         divided: &[AxisAlignedRectangle<T>],
     ) where
@@ -243,14 +413,12 @@ mod tests {
             + std::cmp::PartialOrd,
     {
         // check all divided rectangles are inside the original rectangle
-        for i in 0..divided.len() {
-            assert!(original.enclodes(&divided[i]));
+        for d in divided {
+            assert!(original.enclodes(d));
         }
         // check no overlap between divided rectangles
-        for i in 0..divided.len() {
-            for j in i + 1..divided.len() {
-                assert!(!divided[i].overlaps(&divided[j]));
-            }
+        for (d1, d2) in divided.iter().zip(divided.iter().skip(1)) {
+            assert!(!d1.overlaps(d2));
         }
     }
 }
